Lots of minor cleanups
More comprehensive handling of light calculations. We compute the light To surface vector, then make it bounce on the surface via the normal. We compute a surface To camera vector and the more colinear the bounced light is to the surface To camera vector, the higher the specular will be perceived at this point. For diffuse, we must negate the surface To light because we are taking the dot with the normal.
While those changes add somes unecessary negations in the code, they shouldn't have a high cost.
Also make the shadowMap fetch at the very beginning of the shader to allow the card to hide the fetch latency by processing instructions instead of going idle. Aka Latency Hiding.

The x key toggles the drawing of the scene (objects[0])